+++ src/PSRest/Const.cs
@@ -4,6 +4,7 @@ namespace PSRest;
 static class Const
 {
     public const string EnvNameShared = "$shared";
+    public const string EnvVarDefault = "REST_ENV";
 
     public const string DotEnvFile = ".env";
     public const string SettingsFile = "settings.json";
@@ -14,4 +15,6 @@ static class Const
     public const string VarTypeShared = "$shared";
     public const string VarTypeDotEnv = "$dotenv";
     public const string VarTypeProcessEnv = "$processEnv";
+
+    public const string MediaTypeJson = "application/json";
 }