+++ Release-Notes.md
@@ -1,5 +1,12 @@
 # PSRest Release Notes
 
+## v0.1.0
+
+New cmdlet `Invoke-RestHttp` (.http, .rest) invokes the first HTTP request with
+JSON or GraphQL body with no file variables and no GraphQL variables, for now.
+
+`Set-RestEnvironment` parameter `Name` default is `$env:REST_ENV`.
+
 ## v0.0.3
 
 Commands help and other tweaks.